pyproject.toml
[project]
name = "flaskr"
version = "0.20.4"
description = "The basic blog app built in the Flask tutorial."
dependencies = [
    "flask",
    "flask_sqlalchemy",
    "flask_migrate",
    "faker",
    "psycopg2-binary",
    "flask-wtf",
    "flask-compress",
    "flasgger",
    "markdown==3.7",
    "bleach==6.1.0",
    "pymdown-extensions==10.8.1",
    "beautifulsoup4==4.12.3"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "flake8",
  "flake8-pyproject",
  "bandit[toml]",
  "pip-audit",
  "cyclonedx-bom",
  "pre-commit"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["flaskr"]

[tool.flake8]
indent-size = 2
ignore = ['E121']
max-line-length = 120
count = true
exclude = [
    "__pycache__",
    ".git",
    "migrations/",
    "venv/",
    "build/",
    "dist/",
    ".venv/",
    "node_modules"
]

[tool.bandit]
severity = "medium"
exclude_dirs = ["tests", "migrations", "venv", ".venv", "node_modules"]
